create dummy command if checklist not available
